Create a tmpfs mountpoint at /mnt/disks
This creates a tmpfs mountpoint at /mnt/disks to make sure that a CIFS connection dropout doesn't fill all Unraid RAMFS.
Please note that the tmpfs mountpoint will only be created if no disks are mounted yet, to be sure that all previously mounted disks maintain a valid mountpoint. Because ou that, a reboot may be required.